Cleanup

+++ bing_image_downloader/downloader.py
@@ -1,42 +1,30 @@
-import os, sys
+import sys
 import shutil
 from pathlib import Path
 
 try:
     from bing import Bing
-except ImportError:  # Python 3
+except ImportError:
     from .bing import Bing
 
 
-def download(query, limit=100, output_dir='dataset', adult_filter_off=True, 
-force_replace=False, timeout=60, filter="",resize=None, verbose=True):
+def download(query, limit=100, output_dir='dataset', adult_filter_off=True,
+             force_replace=False, timeout=60, filter='', verbose=True):
 
-    # engine = 'bing'
-    if adult_filter_off:
-        adult = 'off'
-    else:
-        adult = 'on'
+    adult = 'off' if adult_filter_off else 'on'
 
-    
     image_dir = Path(output_dir).joinpath(query).absolute()
 
-    if force_replace:
-        if Path.is_dir(image_dir):
-            shutil.rmtree(image_dir)
+    if force_replace and Path.is_dir(image_dir):
+        shutil.rmtree(image_dir)
 
-    # check directory and create if necessary
     try:
         if not Path.is_dir(image_dir):
             Path.mkdir(image_dir, parents=True)
-
     except Exception as e:
-        print('[Error]Failed to create directory.', e)
+        print('[Error] Failed to create directory.', e)
         sys.exit(1)
-        
-    print("[%] Downloading Images to {}".format(str(image_dir.absolute())))
-    bing = Bing(query, limit, image_dir, adult, timeout, filter,resize, verbose)
+
+    print(f'[%] Downloading Images to {image_dir}')
+    bing = Bing(query, limit, image_dir, adult, timeout, filter, verbose)
     bing.run()
-
-
-if __name__ == '__main__':
-    download('dog', output_dir="..\\Users\\cat", limit=10, timeout=1)
